Bug in memory leak when keep detecting faces for auto-start. Otherwise, everything is functional

+++ MisfitHeartRate/MisfitHeartRate/algorithms/face_params.h
@@ -97,8 +97,8 @@ namespace MHR {
     const int _THRESHOLD_FACE_FRAMES_MIN = 2;
     const int _THRESHOLD_FACE_FRAMES_FOR_START = 5;
     
-    const int _ROI_RATIO_UPPER = 1.5;
-    const int _ROI_RATIO_LOWER = 0.8;
+    const float _ROI_RATIO_UPPER = 1.5f;
+    const float _ROI_RATIO_LOWER = 0.8f;
     
     
 }
